+++ routes/jobRouter.js
@@ -1,20 +1,12 @@
 const jobController = require('../controllers/jobController');
 var router = require('express').Router();
 
-//overview of job offers & form for creating jobs
-// router.get("/jobs", jobController.getAllJobs);
-// router.get("/jobs/new", jobController.createJobs); 
+router.get("/user/:id/offers", jobController.indexJobOffers, jobController.renderJobOffers);
 
-// router.post("/jobs/create", jobController.saveJob);
-
-router.post("/jobs/:jobId/update", jobController.updateJob)
-
-router.get("/jobs/:jobId/delete", jobController.deleteJob);
-
-// edit form
-router.get('/jobs/:jobId/edit', jobController.renderSingleJobEdit);
-
-// render single job
+// single jobs CRUD
 router.get("/jobs/:jobId", jobController.renderSingleJob);
+router.get('/jobs/:jobId/edit', jobController.renderSingleJobEdit);
+router.post("/jobs/:jobId/update", jobController.updateJob)
+router.get("/jobs/:jobId/delete", jobController.deleteJob);
 
 module.exports = router;
